+++ Dave-frontend/src/assets/Accept.png
