+++ Dave-frontend/src/assets/Reject.png
