+++ Dave-frontend/src/assets/patientHistory.jpg
