+++ Dave-frontend/src/assets/symptomToDiagnosis.jpg
