+++ Dave-frontend/src/style/App.css
@@ -8,14 +8,14 @@ body {
   width: 100%;
   display: block;
   position: relative;
-  font-family: 'Gill Sans', 'Gill Sans MT', Calibri, 'Trebuchet MS', sans-serif !important;
+  font-family: "Gill Sans", "Gill Sans MT", Calibri, "Trebuchet MS", sans-serif !important;
 }
 
 #root {
   height: 100%;
   position: relative;
   width: 100%;
-  --primary-clr: #FFFFFF;
+  --primary-clr: #ffffff;
   --graph-btns-clr: #944f96;
   --graph-btns-disabled: hsl(300, 32%, 23%);
   --primary-clr-textbox: #944f96;
@@ -129,7 +129,7 @@ body {
   flex-direction: column;
   justify-content: space-between;
   height: 100%;
-  font-family: 'Gill Sans', 'Gill Sans MT', Calibri, 'Trebuchet MS', sans-serif !important;
+  font-family: "Gill Sans", "Gill Sans MT", Calibri, "Trebuchet MS", sans-serif !important;
 }
 
 .cyto {
@@ -139,7 +139,7 @@ body {
 
 .graphBox-wrapper__clinical-notes__header {
   margin-bottom: 2rem;
-  font-family: 'Gill Sans', 'Gill Sans MT', Calibri, 'Trebuchet MS', sans-serif !important;
+  font-family: "Gill Sans", "Gill Sans MT", Calibri, "Trebuchet MS", sans-serif !important;
 }
 
 #clincalNotesTextField {
@@ -152,11 +152,11 @@ body {
   display: flex;
   flex-direction: row;
   justify-content: space-between;
-  font-family: 'Gill Sans', 'Gill Sans MT', Calibri, 'Trebuchet MS', sans-serif !important;
+  font-family: "Gill Sans", "Gill Sans MT", Calibri, "Trebuchet MS", sans-serif !important;
 }
 
 h5 {
-  font-family: 'Gill Sans', 'Gill Sans MT', Calibri, 'Trebuchet MS', sans-serif !important;
+  font-family: "Gill Sans", "Gill Sans MT", Calibri, "Trebuchet MS", sans-serif !important;
 }
 
 .collapsed-child {
@@ -195,14 +195,14 @@ textarea {
 .graph-box-right__button-container__label {
   align-self: center !important;
   margin-left: 5px !important;
-  font-family: 'Gill Sans', 'Gill Sans MT', Calibri, 'Trebuchet MS', sans-serif !important;
+  font-family: "Gill Sans", "Gill Sans MT", Calibri, "Trebuchet MS", sans-serif !important;
 }
 
 .graph-box-right__button-container__label--disabled {
   align-self: center !important;
   margin-left: 5px !important;
   opacity: 0.3;
-  font-family: 'Gill Sans', 'Gill Sans MT', Calibri, 'Trebuchet MS', sans-serif !important;
+  font-family: "Gill Sans", "Gill Sans MT", Calibri, "Trebuchet MS", sans-serif !important;
 }
 
 .img-box {
@@ -246,11 +246,11 @@ img {
 .btn-class {
   border: 1px solid #944f96 !important;
   color: #944f96 !important;
-  font-family: 'Gill Sans', 'Gill Sans MT', Calibri, 'Trebuchet MS', sans-serif !important;
+  font-family: "Gill Sans", "Gill Sans MT", Calibri, "Trebuchet MS", sans-serif !important;
 }
 
 .epic {
-  background-image: url("./epic-systems-corporation-vector-logo.svg") !important;
+  background-image: url("../assets/epic-systems-corporation-vector-logo.svg") !important;
   background-repeat: no-repeat;
   background-size: 97px 73.5px;
   background-position: center;
@@ -299,7 +299,7 @@ img {
 .legend-label {
   align-self: center !important;
   margin-left: 5px !important;
-  font-family: 'Gill Sans', 'Gill Sans MT', Calibri, 'Trebuchet MS', sans-serif !important;
+  font-family: "Gill Sans", "Gill Sans MT", Calibri, "Trebuchet MS", sans-serif !important;
 }
 
 .legend {
@@ -323,7 +323,7 @@ li::before {
   content: "";
   display: inline-block;
   vertical-align: middle;
-  margin-inline-end: .5rem;
+  margin-inline-end: 0.5rem;
   border-width: 2px;
   border-style: solid;
 }
@@ -364,7 +364,7 @@ li::before {
   width: 1rem;
   height: 1rem;
   border-radius: 3px;
-  transform: rotate(45deg) scale(.85);
+  transform: rotate(45deg) scale(0.85);
 }
 
 .legend-square-consideration::before {
@@ -391,7 +391,7 @@ li::before {
   color: rgb(0, 0, 0, 0.87) !important;
   background-color: #fafafa !important;
   font-weight: normal !important;
-  font-family: 'Gill Sans', 'Gill Sans MT', Calibri, 'Trebuchet MS', sans-serif !important;
+  font-family: "Gill Sans", "Gill Sans MT", Calibri, "Trebuchet MS", sans-serif !important;
   font-size: 13px !important;
   text-align: left !important;
   box-shadow: none !important;
