+++ Dave-frontend/src/style/dialog.css
