+++ Dave-frontend/src/style/equil.css
