--- Dave-frontend/src/styledButtons.js
@@ -1,15 +1,0 @@
-import * as React from 'react';
-import { styled } from '@mui/material/styles';
-import Button from '@mui/material/Button';
-import Stack from '@mui/material/Stack';
-import { purple } from '@mui/material/colors';
-
-const ColorButton = styled(Button)(({ theme }) => ({
-    color: theme.palette.getContrastText(purple[500]),
-    backgroundColor:  '#1E5288',
-    '&:hover': {
-      backgroundColor:  '#1E5288',
-    },
-  }));
-
-export default ColorButton;
