--- bin/python
@@ -1,1 +1,0 @@
-/Users/hadihamoud/opt/anaconda3/bin/python
