--- bin/python3
@@ -1,1 +1,0 @@
-python
