--- bin/python3.9
@@ -1,1 +1,0 @@
-python
