+++ misc./cytotest.py
