+++ misc./medspacy_trials.py
