+++ misc./testingmechanism.py
