--- pyvenv.cfg
@@ -1,8 +1,0 @@
-home = /Users/hadihamoud/opt/anaconda3
-implementation = CPython
-version_info = 3.9.7.final.0
-virtualenv = 20.13.4
-include-system-site-packages = true
-base-prefix = /Users/hadihamoud/opt/anaconda3
-base-exec-prefix = /Users/hadihamoud/opt/anaconda3
-base-executable = /Users/hadihamoud/opt/anaconda3/bin/python
